Please enter the commit message for your changes. Lines starting  with '#' will be ignored, and an empty message aborts the commit.

+++ models/dual_encoder.py
@@ -55,8 +55,12 @@ def dual_encoder_model(
     # Run the utterance and context through the RNN
     rnn_outputs, rnn_states = tf.nn.dynamic_rnn(
         cell,
-        tf.concat(0, [tf.concat(0,[context_embedded,context_token_len_avg, context_nums]),
-                      tf.concat(0,[utterance_embedded,utterance_token_len_avg,utterance_nums])]
+        tf.concat(0, [tf.concat(0,[context_embedded,
+                                   tf.cast(context_token_len_avg,dtype=tf.float32),
+                                   tf.cast(context_nums,dtype=tf.float32)]),
+                      tf.concat(0,[utterance_embedded,
+                                   tf.cast(utterance_token_len_avg,dtype=tf.float32),
+                                   tf.cast(utterance_nums,dtype=tf.float32)])]
                       ),
         sequence_length=tf.concat(0, [context_len+2, utterance_len+2]),
         dtype=tf.float32)
